modified hashing primes to be larger

+++ include/project/primes.h
@@ -3,8 +3,9 @@
 
 #include <math.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 bool is_prime(int num);
-int next_prime(int num);
+int64_t next_prime(int num);
 
 #endif  // HASH_TABLE_PRIMES_H
